requirements.txt
django==4.2.2
djangorestframework==3.14.0
#boto3==1.26.154
drf_yasg==1.21.6
uwsgi